order history completed

+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/payments/checkout/modelPlaceOrder/ModelPlaceOrder.kt
@@ -4,7 +4,7 @@ data class ModelPlaceOrder(
     val __v: Int,
     val _id: String,
     val customer: String,
-    val deliveryCharges: Int,
+    val deliveryCharges: Double,
     val merchant: String,
     val orderType: String,
     val products: List<Product>,
@@ -12,9 +12,9 @@ data class ModelPlaceOrder(
     val shop: String,
     val specialNote: String,
     val status: String,
-    val subTotal: Int,
-    val tax: Int,
-    val total: Int,
+    val subTotal: Double,
+    val tax: Double,
+    val total: Double,
     val useWallet: Boolean,
-    val usedWalletAmount: Int
+    val usedWalletAmount: Double
 )
